+++ templatetags/templatetags/dict.py
@@ -5,7 +5,7 @@ register = template.Library()
 
 @register.filter
 def keyvalue(data, key):
-    if hasattr(data, 'get'):
+    if hasattr(data, "get"):
         return data.get(key)
     else:
         return data[key]