+++ templatetags/templatetags/json.py
@@ -4,6 +4,7 @@ from django import template
 
 register = template.Library()
 
+
 @register.filter
 def pretty_json(value):
     return json.dumps(value, indent=4)